+++ packages/core/tests/Markup.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'rut-dom';
 import Markup from '../src/Markup';
 import Element from '../src/Element';
-import { MOCK_MARKUP } from '../src/testUtils';
+import { MOCK_MARKUP } from '../src/testing';
 import { MarkupProps } from '../src/types';
 
 const options = { log: false, reactElements: false };